pyproject.toml
[tool.poetry]
name = "blog"
version = "0.1.0"
description = ""
authors = ["sreekar998 <sreekarreddy98k@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
fastapi = "^0.116.1"
sqlalchemy = "^2.0.43"
alembic = "^1.16.5"
bcrypt = "^4.3.0"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
pydantic = {extras = ["email"], version = "^2.11.7"}
uvicorn = "^0.35.0"
passlib = "^1.7.4"
python-multipart = "^0.0.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
